[refactor] 좌석 선택 페이지 상태 로직 분리
