+++ src/pages/books/ui/SeatsPage.tsx
@@ -1,167 +1,95 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
-import type { PointerEvent as ReactPointerEvent } from 'react';
+import { useEffect, useMemo, useState } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
 
-import { createSeatsForZone } from '@/pages/books/model/seatData';
-import type { ResellListingItem } from '@/pages/books/model/resellData';
-import { getSelectedSeatDetails } from '@/pages/books/model/selectedSeats';
-import { useSeatHoldActions } from '@/pages/books/model/useSeatHoldActions';
-import { useSeatHoldStore } from '@/entities/seat-hold/model/useSeatHoldStore';
-import { useSeatSelectionStore } from '@/entities/seat-selection/model/useSeatSelectionStore';
-import { useSeatMapData } from '@/pages/books/model/useSeatMapData';
-import { useResellZoneInsights } from '@/pages/books/model/useResellZoneInsights';
-import {
-   holdResaleListing,
-   releaseResaleListingHold,
-   releaseResaleListingHoldKeepalive,
-} from '@/entities/resale/api/resaleApi';
-import { formatPrice, getBookingZones, getZoneOverviewImage, getStadiumName } from '@/pages/books/model/zoneData';
 import type { SeatItem } from '@/pages/books/model/types';
-import { getBookingFlowMode } from '@/shared/lib/booking-flow';
-import { useBookingEntryStore, type BookingEntryState } from '@/shared/lib/useBookingEntryStore';
-import { getErrorMessage } from '@/shared/lib/error/getErrorMessage';
+import { buildResellPaymentEntry } from '@/pages/books/model/buildSeatPaymentEntry';
+import { useResellSeatSelection } from '@/pages/books/model/useResellSeatSelection';
+import { useSeatMapLayout } from '@/pages/books/model/useSeatMapLayout';
+import { useSeatMapViewport } from '@/pages/books/model/useSeatMapViewport';
+import { useSeatsPageEntry } from '@/pages/books/model/useSeatsPageEntry';
+import { useZoneSeatState } from '@/pages/books/model/useZoneSeatState';
+import { formatPrice } from '@/pages/books/model/zoneData';
 import { useBotDetector } from '@/shared/lib/useBotDetector';
 import { Drawer, DrawerContent, DrawerTrigger } from '@/shared/ui/drawer';
-import SeatMapStage from './components/SeatMapStage';
+
 import ResellSeatSidebar from './components/ResellSeatSidebar';
 import ResellZonePreviewSheet from './components/ResellZonePreviewSheet';
+import SeatMapStage from './components/SeatMapStage';
 import SelectedSeatSummaryList from './components/SelectedSeatSummaryList';
-
-const MIN_SCALE = 0.8;
-const MAX_SCALE = 2.4;
-const STAGE_WIDTH = 1240;
-const BLOCK_SEAT_SIZE = 18;
-const BLOCK_SEAT_GAP = 2;
-const BLOCK_CARD_COLUMN_GAP = 48;
-const DEFAULT_SEAT_MAP_LEFT_PADDING = 24;
-const DEFAULT_SEAT_MAP_TOP_PADDING = 24;
-const MINIMAP_WIDTH = 215;
-const MINIMAP_HEIGHT = 140;
-const MINIMAP_PADDING_X = 24;
-const MINIMAP_PADDING_Y = 18;
-
-const normalizeSeatLookupToken = (value: string) => value.replace(/\s+/g, '').toUpperCase();
-
-const parseResellSeatInfo = (seatInfo: string) => {
-   const sectionMatch = seatInfo.match(/([A-Z0-9-]+)구역/i);
-   const rowMatch = seatInfo.match(/([A-Z0-9가-힣]+)열/i);
-   const seatNumberMatch = seatInfo.match(/(\d+)번/);
-
-   if (!sectionMatch || !rowMatch || !seatNumberMatch) {
-      return null;
-   }
-
-   return {
-      sectionCode: normalizeSeatLookupToken(sectionMatch[1]),
-      rowLabel: `${rowMatch[1]}열`,
-      seatNumber: Number(seatNumberMatch[1]),
-   };
-};
 
 function SeatsPage() {
    const navigate = useNavigate();
-   const location = useLocation();
    const { zoneId = '' } = useParams();
-   const bookingFlowMode = getBookingFlowMode(location.search);
-   const isResellMode = bookingFlowMode === 'resell';
-   const routeBookingEntryState = location.state as BookingEntryState | null;
-   const storedBookingEntryState = useBookingEntryStore(state => state.entry);
-   const bookingEntryState = routeBookingEntryState ?? storedBookingEntryState;
-   const setBookingEntry = useBookingEntryStore(state => state.setEntry);
    const { getBotReport } = useBotDetector();
    const botData = getBotReport();
-   const bookingZones = useMemo(
-      () => bookingEntryState?.bookingZones ?? getBookingZones(bookingEntryState?.homeTeamId),
-      [bookingEntryState?.bookingZones, bookingEntryState?.homeTeamId],
-   );
-
-   const zone = useMemo(() => bookingZones.find(item => item.id === zoneId) ?? bookingZones[0], [bookingZones, zoneId]);
-   const zoneOverviewImage = useMemo(
-      () => getZoneOverviewImage(bookingEntryState?.homeTeamId, zone.id),
-      [bookingEntryState?.homeTeamId, zone.id],
-   );
-   const stadiumName = useMemo(() => getStadiumName(bookingEntryState?.homeTeamId), [bookingEntryState?.homeTeamId]);
-
-   const initialSeats = useMemo(() => createSeatsForZone(zone), [zone]);
-   const { apiSeatItems, seatBlocks, hasApiSeatMap, isSeatMapLoading, refetchSeatMap } = useSeatMapData({
-      gameId: bookingEntryState?.gameId,
-      stadiumId: bookingEntryState?.stadiumId,
+   const { bookingEntryState, isResellMode, bookingZones, stadiumName, zone, zoneOverviewImage } = useSeatsPageEntry(zoneId);
+   const {
+      allSelectedSeatsAreHeld: allStandardSelectedSeatsAreHeld,
+      clearSelectedSeats,
+      holdSeat,
+      holdsBySeatId,
+      isSeatInteractionLocked,
+      pendingSeatIds,
+      releaseSeat,
+      seatBlocks,
+      seats,
+      selectedPrice,
+      selectedSeatIds,
+      selectedSeats,
+   } = useZoneSeatState({
+      bookingEntryState,
+      bookingZones,
       zone,
    });
-   const zonesState = useSeatSelectionStore(state => state.zones);
-   const zoneSeatState = useSeatSelectionStore(state => state.zones[zone.id]);
-   const initializeZone = useSeatSelectionStore(state => state.initializeZone);
-   const applyServerSeatSnapshot = useSeatSelectionStore(state => state.applyServerSeatSnapshot);
-   const clearAllSelections = useSeatSelectionStore(state => state.clearAllSelections);
-   const holdsBySeatId = useSeatHoldStore(state => state.holdsBySeatId);
-   const { clearSelectedSeats, holdSeat, pendingSeatIds, releaseSeat, syncHeldSeatsIntoZone } = useSeatHoldActions(
-      bookingEntryState,
-      {
-         onSeatHoldConflict: async () => {
-            await refetchSeatMap();
-         },
-      },
-   );
-
-   const [seatMapScale, setSeatMapScale] = useState(1);
-   const [seatMapOffset, setSeatMapOffset] = useState({ x: 0, y: 0 });
-   const [isSeatMapDragging, setIsSeatMapDragging] = useState(false);
    const [isSeatDrawerOpen, setIsSeatDrawerOpen] = useState(true);
-   const [isResellHoldPending, setIsResellHoldPending] = useState(false);
-   const [selectedResellHoldId, setSelectedResellHoldId] = useState<string | null>(null);
-   const [selectedResellSeatId, setSelectedResellSeatId] = useState<string | null>(null);
-   const dragStartRef = useRef<{ x: number; y: number } | null>(null);
-   const mapViewportRef = useRef<HTMLDivElement | null>(null);
-   const selectedResellHoldIdRef = useRef<string | null>(null);
-   const persistResellHoldRef = useRef(false);
-   const [mapViewportSize, setMapViewportSize] = useState({ width: 0, height: 0 });
-
-   useEffect(() => {
-      if (routeBookingEntryState) {
-         setBookingEntry(routeBookingEntryState);
+   const standardSelectedSeatIdSet = useMemo(() => new Set(selectedSeatIds), [selectedSeatIds]);
+   const resellSeatSelection = useResellSeatSelection({
+      bookingEntryState,
+      isResellMode,
+      isSeatInteractionLocked,
+      seats,
+      zone,
+   });
+   const baseSeatMapLayout = useSeatMapLayout({
+      mapViewportSize: { width: 0, height: 0 },
+      seatBlocks,
+      seatMapOffset: { x: 0, y: 0 },
+      seatMapScale: 1,
+   });
+   const seatMapViewport = useSeatMapViewport({
+      sectionBounds: baseSeatMapLayout.sectionBounds,
+      zoneId: zone.id,
+   });
+   const seatMapLayout = useSeatMapLayout({
+      mapViewportSize: seatMapViewport.mapViewportSize,
+      seatBlocks,
+      seatMapOffset: seatMapViewport.seatMapOffset,
+      seatMapScale: seatMapViewport.seatMapScale,
+   });
+   const selectedSeatIdSet = isResellMode ? resellSeatSelection.selectedSeatIdSet : standardSelectedSeatIdSet;
+   const selectedSeatCount = isResellMode ? resellSeatSelection.selectedSeatCount : selectedSeats.length;
+   const summaryPrice = isResellMode ? resellSeatSelection.summaryPrice : selectedPrice;
+   const allSelectedSeatsAreHeld = isResellMode
+      ? Boolean(resellSeatSelection.selectedResellListing) && Boolean(resellSeatSelection.selectedResellHoldId)
+      : allStandardSelectedSeatsAreHeld;
+   const bookingButtonLabel = isResellMode ? '예매하기' : `${selectedSeats.length}매 예매하기`;
+   const displaySeats = useMemo(() => {
+      if (isResellMode) {
+         return resellSeatSelection.displaySeats;
       }
-   }, [routeBookingEntryState, setBookingEntry]);
 
-   useEffect(() => {
-      const syncedInitialSeats = syncHeldSeatsIntoZone(zone.id, initialSeats);
-
-      if (hasApiSeatMap && apiSeatItems.length > 0) {
-         applyServerSeatSnapshot(zone.id, syncHeldSeatsIntoZone(zone.id, apiSeatItems));
-         return;
+      if (isSeatInteractionLocked) {
+         return seats.map(
+            seat =>
+               ({
+                  ...seat,
+                  status: 'disabled',
+               }) satisfies SeatItem,
+         );
       }
 
-      initializeZone(zone.id, syncedInitialSeats);
-   }, [
-      apiSeatItems,
-      applyServerSeatSnapshot,
-      hasApiSeatMap,
-      initialSeats,
-      initializeZone,
-      syncHeldSeatsIntoZone,
-      zone.id,
-   ]);
-
-   useEffect(() => {
-      const updateViewportSize = () => {
-         const element = mapViewportRef.current;
-
-         if (!element) {
-            return;
-         }
-
-         setMapViewportSize({
-            width: element.clientWidth,
-            height: element.clientHeight,
-         });
-      };
-
-      updateViewportSize();
-      window.addEventListener('resize', updateViewportSize);
-
-      return () => {
-         window.removeEventListener('resize', updateViewportSize);
-      };
-   }, []);
+      return seats;
+   }, [isResellMode, isSeatInteractionLocked, resellSeatSelection.displaySeats, seats]);
 
    useEffect(() => {
       const mediaQuery = window.matchMedia('(min-width: 1280px)');
@@ -179,246 +107,20 @@ function SeatsPage() {
       };
    }, []);
 
-   useEffect(() => {
-      if (!isResellMode) {
-         setSelectedResellHoldId(null);
-         setSelectedResellSeatId(null);
-         persistResellHoldRef.current = false;
-         return;
-      }
-
-      clearAllSelections();
-   }, [clearAllSelections, isResellMode, zone.id]);
-
-   useEffect(() => {
-      selectedResellHoldIdRef.current = selectedResellHoldId;
-   }, [selectedResellHoldId]);
-
-   useEffect(() => {
-      if (!isResellMode) {
-         return;
-      }
-
-      const handlePageHide = () => {
-         const holdId = selectedResellHoldIdRef.current;
-
-         if (!holdId || persistResellHoldRef.current) {
-            return;
-         }
-
-         releaseResaleListingHoldKeepalive(holdId);
-      };
-
-      window.addEventListener('pagehide', handlePageHide);
-
-      return () => {
-         window.removeEventListener('pagehide', handlePageHide);
-      };
-   }, [isResellMode]);
-
-   useEffect(() => {
-      return () => {
-         const holdId = selectedResellHoldIdRef.current;
-
-         if (!isResellMode || !holdId || persistResellHoldRef.current) {
-            return;
-         }
-
-         releaseResaleListingHoldKeepalive(holdId);
-      };
-   }, [isResellMode]);
-
-   const seats = useMemo(() => {
-      if (!zoneSeatState) {
-         return initialSeats;
-      }
-
-      return zoneSeatState.seatOrder
-         .map(seatId => zoneSeatState.seatMap[seatId])
-         .filter((seat): seat is SeatItem => Boolean(seat));
-   }, [initialSeats, zoneSeatState]);
-
-   const selectedSeatIds = zoneSeatState?.selectedSeatIds ?? [];
-   const selectedSeatIdSet = useMemo(
-      () => new Set(isResellMode ? (selectedResellSeatId ? [selectedResellSeatId] : []) : selectedSeatIds),
-      [isResellMode, selectedResellSeatId, selectedSeatIds],
-   );
-
-   const selectedSeats = useMemo(() => getSelectedSeatDetails(zonesState, bookingZones), [bookingZones, zonesState]);
-
-   const selectedPrice = selectedSeats.reduce((total, item) => total + item.price, 0);
-   const isSeatInteractionLocked = Boolean(bookingEntryState?.gameId) && isSeatMapLoading;
-   const resellInsightsQuery = useResellZoneInsights({
-      enabled: isResellMode,
-      gameId: bookingEntryState?.gameId,
-      zone,
-      seats,
-   });
-   const resellInsights = resellInsightsQuery.data ?? null;
-   const resellListingBySeatId = useMemo(() => {
-      const listingBySeatIdentifier = new Map(
-         (resellInsights?.listings ?? []).map(listing => [listing.seatId, listing]),
-      );
-      const seatByLookupKey = new Map(
-         seats.map(seat => [`${normalizeSeatLookupToken(seat.block)}::${seat.rowLabel}::${seat.seatNumber}`, seat]),
-      );
-      const nextMap = new Map<string, ResellListingItem>();
-
-      seats.forEach(seat => {
-         const matchedListing = listingBySeatIdentifier.get(seat.id) ?? listingBySeatIdentifier.get(seat.apiSeatId);
-
-         if (matchedListing) {
-            nextMap.set(seat.id, matchedListing);
-         }
-      });
-
-      (resellInsights?.listings ?? []).forEach(listing => {
-         if ([...nextMap.values()].some(mappedListing => mappedListing.listingId === listing.listingId)) {
-            return;
-         }
-
-         const parsedSeatInfo = parseResellSeatInfo(listing.seatInfo);
-
-         if (!parsedSeatInfo) {
-            return;
-         }
-
-         const matchedSeat = seatByLookupKey.get(
-            `${parsedSeatInfo.sectionCode}::${parsedSeatInfo.rowLabel}::${parsedSeatInfo.seatNumber}`,
-         );
-
-         if (matchedSeat) {
-            nextMap.set(matchedSeat.id, listing);
-         }
-      });
-
-      return nextMap;
-   }, [resellInsights?.listings, seats]);
-   const seatIdByResellListingId = useMemo(() => {
-      return new Map(
-         [...resellListingBySeatId.entries()].map(([seatId, listing]) => [listing.listingId, seatId] as const),
-      );
-   }, [resellListingBySeatId]);
-   const selectedResellListing = useMemo(() => {
-      if (!isResellMode) {
-         return null;
-      }
-
-      return selectedResellSeatId ? (resellListingBySeatId.get(selectedResellSeatId) ?? null) : null;
-   }, [isResellMode, resellListingBySeatId, selectedResellSeatId]);
-   const selectedSeatCount = isResellMode ? (selectedResellListing ? 1 : 0) : selectedSeats.length;
-   const summaryPrice = isResellMode
-      ? (selectedResellListing?.totalAmount ?? selectedResellListing?.listingPrice ?? 0)
-      : selectedPrice;
-   const allSelectedSeatsAreHeld = isResellMode
-      ? Boolean(selectedResellSeatId) && Boolean(selectedResellHoldId)
-      : selectedSeats.every(selectedSeat => Boolean(holdsBySeatId[selectedSeat.seat.id]?.holdId));
-   const bookingButtonLabel = isResellMode ? '예매하기' : `${selectedSeats.length}매 예매하기`;
-   const displaySeats = useMemo(() => {
-      if (isSeatInteractionLocked) {
-         return seats.map(
-            seat =>
-               ({
-                  ...seat,
-                  status: 'disabled',
-               }) satisfies SeatItem,
-         );
-      }
-
-      if (!isResellMode) {
-         return seats;
-      }
-
-      return seats.map(seat => {
-         if (resellListingBySeatId.has(seat.id)) {
-            return {
-               ...seat,
-               status: selectedSeatIdSet.has(seat.id) ? 'selected' : 'available',
-            } satisfies SeatItem;
-         }
-
-         return {
-            ...seat,
-            status: 'disabled',
-         } satisfies SeatItem;
-      });
-   }, [isResellMode, isSeatInteractionLocked, resellListingBySeatId, seats, selectedSeatIdSet]);
-
-   const releaseSelectedResellHold = async () => {
-      if (!selectedResellHoldIdRef.current) {
-         setSelectedResellHoldId(null);
-         setSelectedResellSeatId(null);
-         return;
-      }
-
-      const holdId = selectedResellHoldIdRef.current;
-
-      await releaseResaleListingHold(holdId);
-      selectedResellHoldIdRef.current = null;
-      setSelectedResellHoldId(null);
-      setSelectedResellSeatId(null);
-   };
-
-   const handleSelectResellListing = async (seatId: string, listing: ResellListingItem) => {
-      if (!bookingEntryState?.queueTokenJti || isResellHoldPending) {
-         return;
-      }
-
-      setIsResellHoldPending(true);
-
-      try {
-         const isAlreadySelected = selectedResellSeatId === seatId;
-
-         if (isAlreadySelected && selectedResellHoldIdRef.current) {
-            await releaseSelectedResellHold();
-            return;
-         }
-
-         if (selectedResellHoldIdRef.current) {
-            await releaseSelectedResellHold();
-         }
-
-         const hold = await holdResaleListing({
-            listingId: listing.listingId,
-            queueTokenJti: bookingEntryState.queueTokenJti,
-         });
-
-         selectedResellHoldIdRef.current = hold.holdId;
-         setSelectedResellHoldId(hold.holdId);
-         setSelectedResellSeatId(seatId);
-      } catch (error) {
-         window.alert(getErrorMessage(error, '리셀 좌석 점유 중 오류가 발생했습니다.'));
-      } finally {
-         setIsResellHoldPending(false);
-      }
-   };
-
    const handleProceedToPayment = () => {
       if (isResellMode) {
-         if (!selectedResellListing) {
+         if (!resellSeatSelection.selectedResellListing) {
             return;
          }
 
-         persistResellHoldRef.current = true;
-
+         resellSeatSelection.persistResellHoldRef.current = true;
          navigate('/tickets/resell-payment', {
-            state: {
-               buyerId: bookingEntryState?.userId,
-               listingId: selectedResellListing.listingId,
-               holdId: selectedResellHoldId,
-               queueTokenJti: bookingEntryState?.queueTokenJti,
-               sellerId: selectedResellListing.sellerId,
-               settlementAmount: selectedResellListing.settlementAmount,
-               totalAmount: selectedResellListing.totalAmount,
-               totalBuyerFee: selectedResellListing.totalBuyerFee,
-               totalSellerFee: selectedResellListing.totalSellerFee,
-               seatInfo: selectedResellListing.seatInfo,
-               matchTitle: bookingEntryState?.matchTitle,
-               venue: bookingEntryState?.venue,
-               dateTime: bookingEntryState?.dateTime,
-               turnstileToken: bookingEntryState?.turnstileToken,
+            state: buildResellPaymentEntry({
+               bookingEntryState,
                botData,
-            },
+               holdId: resellSeatSelection.selectedResellHoldId,
+               listing: resellSeatSelection.selectedResellListing,
+            }),
          });
          return;
       }
@@ -428,173 +130,12 @@ function SeatsPage() {
       });
    };
 
-   const sectionBounds = useMemo(() => {
-      if (seatBlocks.length === 0) {
-         return null;
-      }
-
-      const blockMetrics = seatBlocks.map((block, index) => {
-         const width = block.cols * (BLOCK_SEAT_SIZE + BLOCK_SEAT_GAP) - BLOCK_SEAT_GAP;
-         const height = block.rows * (BLOCK_SEAT_SIZE + BLOCK_SEAT_GAP) - BLOCK_SEAT_GAP;
-         const renderedOffsetX = block.offsetX + index * BLOCK_CARD_COLUMN_GAP;
-
-         return {
-            ...block,
-            width,
-            height,
-            renderedOffsetX,
-            right: renderedOffsetX + width,
-            bottom: block.offsetY + height,
-         };
-      });
-
-      return {
-         left: Math.min(...blockMetrics.map(block => block.renderedOffsetX)),
-         top: Math.min(...blockMetrics.map(block => block.offsetY)),
-         right: Math.max(...blockMetrics.map(block => block.right)),
-         bottom: Math.max(...blockMetrics.map(block => block.bottom)),
-         blocks: blockMetrics,
-      };
-   }, [seatBlocks]);
-
-   const directionBadgePosition = useMemo(() => {
-      if (!sectionBounds) {
-         return null;
-      }
-
-      const badgeHeight = 48;
-      const sectionHeight = sectionBounds.bottom - sectionBounds.top;
-      const badgeOffset = Math.max(156, Math.min(260, Math.round(sectionHeight * 0.44)));
-
-      return {
-         left: (sectionBounds.left + sectionBounds.right) / 2,
-         top: Math.max(-160, sectionBounds.top - badgeHeight - badgeOffset),
-      };
-   }, [sectionBounds]);
-
-   const minimapLayout = useMemo(() => {
-      if (!sectionBounds) {
-         return null;
-      }
-
-      const width = sectionBounds.right - sectionBounds.left;
-      const height = sectionBounds.bottom - sectionBounds.top;
-      const scale = Math.min(
-         (MINIMAP_WIDTH - MINIMAP_PADDING_X * 2) / width,
-         (MINIMAP_HEIGHT - MINIMAP_PADDING_Y * 2) / height,
-      );
-      const offsetX = (MINIMAP_WIDTH - width * scale) / 2;
-      const offsetY = (MINIMAP_HEIGHT - height * scale) / 2;
-
-      return {
-         bounds: sectionBounds,
-         scale,
-         offsetX,
-         offsetY,
-         blocks: sectionBounds.blocks.map(block => ({
-            id: block.id,
-            x: offsetX + (block.renderedOffsetX - sectionBounds.left) * scale,
-            y: offsetY + (block.offsetY - sectionBounds.top) * scale,
-            width: block.width * scale,
-            height: block.height * scale,
-         })),
-      };
-   }, [sectionBounds]);
-
-   const minimapViewport = useMemo(() => {
-      if (!sectionBounds) {
-         return null;
-      }
-
-      if (mapViewportSize.width === 0 || mapViewportSize.height === 0) {
-         return null;
-      }
-
-      const stageLeft = (mapViewportSize.width - STAGE_WIDTH * seatMapScale) / 2 + seatMapOffset.x;
-      const stageTop = 56 + seatMapOffset.y;
-      const visibleLeft = (0 - stageLeft) / seatMapScale;
-      const visibleTop = (0 - stageTop) / seatMapScale;
-      const visibleRight = (mapViewportSize.width - stageLeft) / seatMapScale;
-      const visibleBottom = (mapViewportSize.height - stageTop) / seatMapScale;
-
-      const intersectLeft = Math.max(sectionBounds.left, visibleLeft);
-      const intersectTop = Math.max(sectionBounds.top, visibleTop);
-      const intersectRight = Math.min(sectionBounds.right, visibleRight);
-      const intersectBottom = Math.min(sectionBounds.bottom, visibleBottom);
-
-      if (intersectRight <= intersectLeft || intersectBottom <= intersectTop || !minimapLayout) {
-         return null;
-      }
-
-      return {
-         left: minimapLayout.offsetX + (intersectLeft - sectionBounds.left) * minimapLayout.scale,
-         top: minimapLayout.offsetY + (intersectTop - sectionBounds.top) * minimapLayout.scale,
-         width: (intersectRight - intersectLeft) * minimapLayout.scale,
-         height: (intersectBottom - intersectTop) * minimapLayout.scale,
-      };
-   }, [
-      mapViewportSize.height,
-      mapViewportSize.width,
-      minimapLayout,
-      seatMapOffset.x,
-      seatMapOffset.y,
-      seatMapScale,
-      sectionBounds,
-   ]);
-
-   const getDefaultSeatMapView = () => {
-      if (!sectionBounds || mapViewportSize.width === 0 || mapViewportSize.height === 0) {
-         return {
-            scale: 1,
-            offset: { x: 0, y: 0 },
-         };
-      }
-
-      const contentWidth = sectionBounds.right - sectionBounds.left;
-      const contentHeight = sectionBounds.bottom - sectionBounds.top;
-      const availableWidth = Math.max(1, mapViewportSize.width - DEFAULT_SEAT_MAP_LEFT_PADDING * 2);
-      const availableHeight = Math.max(1, mapViewportSize.height - 56 - DEFAULT_SEAT_MAP_TOP_PADDING * 2);
-      const scale = Math.min(
-         MAX_SCALE,
-         Math.max(MIN_SCALE, Math.min(availableWidth / contentWidth, availableHeight / contentHeight)),
-      );
-      const contentCenterX = (sectionBounds.left + sectionBounds.right) / 2;
-      const contentCenterY = (sectionBounds.top + sectionBounds.bottom) / 2;
-      const targetCenterY = 56 + DEFAULT_SEAT_MAP_TOP_PADDING + availableHeight / 2;
-
-      return {
-         scale: Number(scale.toFixed(2)),
-         offset: {
-            x: (STAGE_WIDTH / 2 - contentCenterX) * scale,
-            y: targetCenterY - 56 - contentCenterY * scale,
-         },
-      };
-   };
-
-   const updateSeatMapScale = (nextScale: number) => {
-      setSeatMapScale(Math.min(MAX_SCALE, Math.max(MIN_SCALE, Number(nextScale.toFixed(2)))));
-   };
-
-   const resetSeatMapView = () => {
-      const nextView = getDefaultSeatMapView();
-
-      setSeatMapScale(nextView.scale);
-      setSeatMapOffset(nextView.offset);
-   };
-
-   useEffect(() => {
-      const nextView = getDefaultSeatMapView();
-
-      setSeatMapScale(nextView.scale);
-      setSeatMapOffset(nextView.offset);
-   }, [mapViewportSize.height, mapViewportSize.width, sectionBounds, zone.id]);
-
    const toggleSeat = (seat: SeatItem) => {
       if (isSeatInteractionLocked) {
          return;
       }
 
-      if (pendingSeatIds.includes(seat.id) || isResellHoldPending) {
+      if (pendingSeatIds.includes(seat.id) || resellSeatSelection.isResellHoldPending) {
          return;
       }
 
@@ -603,13 +144,17 @@ function SeatsPage() {
       }
 
       if (isResellMode) {
-         const listing = resellListingBySeatId.get(seat.id);
+         const listing = resellSeatSelection.seatIdByResellListingId
+            ? resellSeatSelection.resellInsights?.listings.find(
+                 item => resellSeatSelection.seatIdByResellListingId.get(item.listingId) === seat.id,
+              )
+            : null;
 
          if (!listing) {
             return;
          }
 
-         void handleSelectResellListing(seat.id, listing);
+         void resellSeatSelection.handleSelectResellListing(seat.id, listing);
          return;
       }
 
@@ -618,7 +163,7 @@ function SeatsPage() {
 
    const handleRemoveSelectedSeat = (selectedZoneId: string, seatId: string) => {
       if (isResellMode) {
-         void releaseSelectedResellHold();
+         void resellSeatSelection.releaseSelectedResellHold();
          return;
       }
 
@@ -633,45 +178,11 @@ function SeatsPage() {
 
    const handleClearSelectedSeats = () => {
       if (isResellMode) {
-         void releaseSelectedResellHold();
+         void resellSeatSelection.releaseSelectedResellHold();
          return;
       }
 
       void clearSelectedSeats(selectedSeats);
-   };
-
-   const handleMapPointerDown = (event: ReactPointerEvent<HTMLDivElement>) => {
-      const target = event.target as HTMLElement;
-      if (target.closest('button')) {
-         return;
-      }
-
-      dragStartRef.current = {
-         x: event.clientX - seatMapOffset.x,
-         y: event.clientY - seatMapOffset.y,
-      };
-      setIsSeatMapDragging(true);
-      event.currentTarget.setPointerCapture(event.pointerId);
-   };
-
-   const handleMapPointerMove = (event: ReactPointerEvent<HTMLDivElement>) => {
-      if (!isSeatMapDragging || !dragStartRef.current) {
-         return;
-      }
-
-      setSeatMapOffset({
-         x: event.clientX - dragStartRef.current.x,
-         y: event.clientY - dragStartRef.current.y,
-      });
-   };
-
-   const handleMapPointerUp = (event: ReactPointerEvent<HTMLDivElement>) => {
-      dragStartRef.current = null;
-      setIsSeatMapDragging(false);
-
-      if (event.currentTarget.hasPointerCapture(event.pointerId)) {
-         event.currentTarget.releasePointerCapture(event.pointerId);
-      }
    };
 
    return (
@@ -689,24 +200,24 @@ function SeatsPage() {
 
                <div className="relative flex-1 overflow-hidden px-0 pb-[144px] lg:px-8 lg:pb-6 xl:pb-6">
                   <SeatMapStage
-                     directionBadgePosition={directionBadgePosition}
-                     isSeatMapDragging={isSeatMapDragging}
-                     mapViewportRef={mapViewportRef}
-                     minimapLayout={minimapLayout}
-                     minimapViewport={minimapViewport}
+                     directionBadgePosition={seatMapLayout.directionBadgePosition}
+                     isSeatMapDragging={seatMapViewport.isSeatMapDragging}
+                     mapViewportRef={seatMapViewport.mapViewportRef}
+                     minimapLayout={seatMapLayout.minimapLayout}
+                     minimapViewport={seatMapLayout.minimapViewport}
                      seatBlocks={seatBlocks}
-                     seatMapOffset={seatMapOffset}
-                     seatMapScale={seatMapScale}
+                     seatMapOffset={seatMapViewport.seatMapOffset}
+                     seatMapScale={seatMapViewport.seatMapScale}
                      seats={displaySeats}
                      selectedSeatIdSet={selectedSeatIdSet}
                      zoneColor={zone.color}
                      zoneName={zone.name}
-                     onMapPointerDown={handleMapPointerDown}
-                     onMapPointerMove={handleMapPointerMove}
-                     onMapPointerUp={handleMapPointerUp}
-                     onResetSeatMapView={resetSeatMapView}
+                     onMapPointerDown={seatMapViewport.onMapPointerDown}
+                     onMapPointerMove={seatMapViewport.onMapPointerMove}
+                     onMapPointerUp={seatMapViewport.onMapPointerUp}
+                     onResetSeatMapView={seatMapViewport.resetSeatMapView}
                      onToggleSeat={toggleSeat}
-                     onUpdateSeatMapScale={updateSeatMapScale}
+                     onUpdateSeatMapScale={seatMapViewport.onUpdateSeatMapScale}
                   />
                </div>
 
@@ -743,30 +254,30 @@ function SeatsPage() {
                      className="overflow-hidden border-none p-0 xl:hidden"
                   >
                      <div className="h-full overflow-y-auto">
-                        {isResellMode && resellInsightsQuery.isPending ? (
+                        {isResellMode && resellSeatSelection.resellInsightsQuery.isPending ? (
                            <div className="flex h-full min-h-[240px] items-center justify-center px-5 text-center text-body-1-medium text-muted-foreground">
                               리셀 좌석 정보를 불러오는 중입니다.
                            </div>
-                        ) : isResellMode && resellInsightsQuery.isError ? (
+                        ) : isResellMode && resellSeatSelection.resellInsightsQuery.isError ? (
                            <div className="flex h-full min-h-[240px] items-center justify-center px-5 text-center text-body-1-medium text-muted-foreground">
                               리셀 좌석 정보를 불러오지 못했습니다.
                            </div>
-                        ) : isResellMode && resellInsights ? (
+                        ) : isResellMode && resellSeatSelection.resellInsights ? (
                            <ResellZonePreviewSheet
-                              insights={resellInsights}
+                              insights={resellSeatSelection.resellInsights}
                               zone={zone}
-                              selectedListingId={selectedResellListing?.listingId}
+                              selectedListingId={resellSeatSelection.selectedResellListing?.listingId}
                               onSelectListing={listing => {
-                                 const mappedSeatId = seatIdByResellListingId.get(listing.listingId);
+                                 const mappedSeatId = resellSeatSelection.seatIdByResellListingId.get(listing.listingId);
 
                                  if (!mappedSeatId) {
                                     return;
                                  }
 
-                                 void handleSelectResellListing(mappedSeatId, listing);
+                                 void resellSeatSelection.handleSelectResellListing(mappedSeatId, listing);
                               }}
                               submitLabel={bookingButtonLabel}
-                              submitDisabled={!selectedResellListing}
+                              submitDisabled={!resellSeatSelection.selectedResellListing}
                               onSubmit={handleProceedToPayment}
                            />
                         ) : (
@@ -823,29 +334,29 @@ function SeatsPage() {
                </Drawer>
             </section>
 
-            {isResellMode && resellInsightsQuery.isPending ? (
+            {isResellMode && resellSeatSelection.resellInsightsQuery.isPending ? (
                <aside className="hidden w-full shrink-0 items-center justify-center border-l border-border-light bg-background px-5 text-center text-body-1-medium text-muted-foreground xl:flex xl:w-[420px]">
                   리셀 좌석 정보를 불러오는 중입니다.
                </aside>
-            ) : isResellMode && resellInsights ? (
+            ) : isResellMode && resellSeatSelection.resellInsights ? (
                <ResellSeatSidebar
-                  insights={resellInsights}
-                  selectedListingId={selectedResellListing?.listingId}
+                  insights={resellSeatSelection.resellInsights}
+                  selectedListingId={resellSeatSelection.selectedResellListing?.listingId}
                   zone={zone}
                   zoneOverviewImage={zoneOverviewImage}
                   stadiumName={stadiumName}
                   onSelectListing={listing => {
-                     const mappedSeatId = seatIdByResellListingId.get(listing.listingId);
+                     const mappedSeatId = resellSeatSelection.seatIdByResellListingId.get(listing.listingId);
 
                      if (!mappedSeatId) {
                         return;
                      }
 
-                     void handleSelectResellListing(mappedSeatId, listing);
+                     void resellSeatSelection.handleSelectResellListing(mappedSeatId, listing);
                   }}
                   onSubmit={handleProceedToPayment}
                />
-            ) : isResellMode && resellInsightsQuery.isError ? (
+            ) : isResellMode && resellSeatSelection.resellInsightsQuery.isError ? (
                <aside className="hidden w-full shrink-0 items-center justify-center border-l border-border-light bg-background px-5 text-center text-body-1-medium text-muted-foreground xl:flex xl:w-[420px]">
                   리셀 좌석 정보를 불러오지 못했습니다.
                </aside>
